Fix bugs with widgets and change scale to use doubles to avoid floating point errors.

+++ src/main/java/com/intro/client/render/drawables/ArmorDisplay.java
@@ -136,7 +136,7 @@ public class ArmorDisplay extends Scalable {
     }
 
     @Override
-    public void onScaleChange(float oldScale, float newScale) {
+    public void onScaleChange(double oldScale, double newScale) {
         OsmiumClient.options.put(Options.ArmorDisplayPosition, new ElementPositionOption(Options.ArmorDisplayPosition, this.posX, this.posY, newScale));
     }
 }
